nasm backend

+++ COMMON/include/helpingFunctions.h
@@ -1,21 +1,22 @@
 #ifndef HELPING_FUNCTIONS_H
 #define HELPING_FUNCTIONS_H
 
-    char* copyFileContent (const char* nameOfFile);
+char* copyFileContent (const char* nameOfFile);
 
-    unsigned int getSizeOfFile (int fileDescriptor);
+long int getSizeOfFile (int fileDescriptor);
 
-    void skipSpaces(char** bufPos, size_t* curLine);
+void skipSpaces (char** bufPos, size_t* curLine);
 
-    int bsearchHashComparator(const void* firstParam, const void* secondParam);
+int bsearchHashComparator (const void* firstParam, const void* secondParam);
 
-    unsigned long long getStringHash (const char* string);
+unsigned long long getStringHash (const char* string);
 
-    int structIdentifierComparator (const void* firstStruct, const void* secondStruct);
+int structIdentifierComparator (const void* firstStruct, const void* secondStruct);
 
-    int compareDouble (double first, double second);
+int compareDouble (double first, double second);
 
-    size_t getSize_t (void);
+size_t getSize_t (void);
 
-    unsigned long long getFactorial(size_t num);
+unsigned long long getFactorial (size_t num);
+
 #endif
